refactor : user, space, pay dto 계층구조로 분리
