+++ src/main/java/space/space_spring/controller/UserController.java
@@ -7,13 +7,13 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import space.space_spring.argument_resolver.jwtLogin.JwtLoginAuth;
-import space.space_spring.dto.user.*;
+import space.space_spring.dto.user.request.PostUserLoginRequest;
+import space.space_spring.dto.user.request.PostUserSignupRequest;
+import space.space_spring.dto.user.response.GetSpaceInfoForUserResponse;
 import space.space_spring.exception.UserException;
 import space.space_spring.response.BaseResponse;
 import space.space_spring.service.UserService;
 import space.space_spring.util.userSpace.UserSpaceUtils;
-
-import java.util.List;
 
 import static space.space_spring.response.status.BaseExceptionResponseStatus.*;
 import static space.space_spring.util.bindingResult.BindingResultUtils.getErrorMessage;
