+++ src/main/java/space/space_spring/dao/SpaceDao.java
@@ -4,11 +4,6 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.stereotype.Repository;
 import space.space_spring.entity.Space;
-import space.space_spring.entity.User;
-import space.space_spring.entity.UserSpace;
-import space.space_spring.dto.space.PostSpaceCreateRequest;
-
-import static space.space_spring.entity.enumStatus.UserSpaceAuth.MANAGER;
 
 @Repository
 public class SpaceDao {
