+++ src/main/java/space/space_spring/dao/UserDao.java
@@ -6,7 +6,6 @@ import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
 import org.springframework.stereotype.Repository;
 import space.space_spring.entity.User;
-import space.space_spring.dto.user.PostUserSignupRequest;
 import space.space_spring.entity.enumStatus.UserSignupType;
 
 @Repository
