+++ src/main/java/space/space_spring/dao/UserSpaceDao.java
@@ -4,9 +4,8 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
 import org.springframework.stereotype.Repository;
-import space.space_spring.dto.user.GetSpaceInfoForUserResponse;
-import space.space_spring.dto.user.SpaceChoiceInfo;
-import space.space_spring.dto.user.SpaceChoiceViewDto;
+import space.space_spring.dto.user.dto.SpaceChoiceInfo;
+import space.space_spring.dto.user.dto.SpaceChoiceViewDto;
 import space.space_spring.dto.userSpace.UserProfileImgAndNameDto;
 import space.space_spring.entity.Space;
 import space.space_spring.entity.User;
