+++ src/main/java/space/space_spring/dto/pay/response/GetPayViewResponse.java
@@ -1,7 +1,9 @@
-package space.space_spring.dto.pay;
+package space.space_spring.dto.pay.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import space.space_spring.dto.pay.dto.PayReceiveInfoDto;
+import space.space_spring.dto.pay.dto.PayRequestInfoDto;
 
 import java.util.ArrayList;
 import java.util.List;
