+++ src/main/java/space/space_spring/dto/pay/response/GetRecentPayRequestBankInfoResponse.java
@@ -1,7 +1,8 @@
-package space.space_spring.dto.pay;
+package space.space_spring.dto.pay.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import space.space_spring.dto.pay.dto.RecentPayRequestBankInfoDto;
 
 import java.util.ArrayList;
 import java.util.List;
