+++ src/main/java/space/space_spring/service/UserService.java
@@ -5,7 +5,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import space.space_spring.dao.UserSpaceDao;
-import space.space_spring.dto.user.*;
+import space.space_spring.dto.user.dto.SpaceChoiceViewDto;
+import space.space_spring.dto.user.request.PostUserLoginRequest;
+import space.space_spring.dto.user.request.PostUserSignupRequest;
+import space.space_spring.dto.user.response.GetSpaceInfoForUserResponse;
 import space.space_spring.entity.enumStatus.UserSignupType;
 import space.space_spring.jwt.JwtLoginProvider;
 import space.space_spring.dao.UserDao;
